AnalyticsMS/requirements.txt
paho-mqtt
Flask==3.0.2
numpy
requests